// src/common/enums/availability.enum.ts
export enum Availability {
  MON_9AM_12PM = 'Mon 9 AM - 12 PM',
  MON_1PM_4PM = 'Mon 1 PM - 4 PM',
  TUE_9AM_12PM = 'Tue 9 AM - 12 PM',
  TUE_1PM_4PM = 'Tue 1 PM - 4 PM',
  WED_9AM_12PM = 'Wed 9 AM - 12 PM',
  WED_1PM_4PM = 'Wed 1 PM - 4 PM',
  THU_9AM_12PM = 'Thu 9 AM - 12 PM',
  THU_1PM_4PM = 'Thu 1 PM - 4 PM',
  FRI_9AM_12PM = 'Fri 9 AM - 12 PM',
  FRI_1PM_4PM = 'Fri 1 PM - 4 PM',
  SAT_9AM_12PM = 'Sat 9 AM - 12 PM',
  SAT_1PM_4PM = 'Sat 1 PM - 4 PM',
  SUN_9AM_12PM = 'Sun 9 AM - 12 PM',
  SUN_1PM_4PM = 'Sun 1 PM - 4 PM',
}