// src/common/enums/city.enum.ts
export enum City {
  ISLAMABAD = 'Islamabad',
  KARACHI = 'Karachi',
  LAHORE = 'Lahore',
  RAWALPINDI = 'Rawalpindi',
  PESHAWAR = 'Peshawar',
  QUETTA = 'Quetta',
  MULTAN = 'Multan',
  FSD = 'Faisalabad',
  SIALKOT = 'Sialkot',
  GUJRANWALA = 'Gujranwala',
  ISLAMABAD_RURAL = 'Islamabad Rural',
  SARGODHA = 'Sargodha',
  ABBOTTABAD = 'Abbottabad',
  BHAKKAR = 'Bhakkar',
  SUKKUR = 'Sukkur',
  MURREE = 'Murree',
  HYDERABAD = 'Hyderabad',
  JHELUM = 'Jhelum',
  GILGIT = 'Gilgit',
  SKARDU = 'Skardu',
  BAHAWALPUR = 'Bahawalpur',
  OKARA = 'Okara',
  CHINIOT = 'Chiniot',
  DERA_GHAZI_KHAN = 'Dera Ghazi Khan',
  CHAKWAL = 'Chakwal',
  MIRPUR = 'Mirpur',
  KOHAT = 'Kohat',
  NAROWAL = 'Narowal',
  MARDAN = 'Mardan',
  SWAT = 'Swat',
  NASEERABAD = 'Naseerabad',
}